integration/requirements.txt
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.0
gevent==23.9.1
greenlet==3.0.1
idna==3.4
iniconfig==2.0.0
ntplib==0.4.0
packaging==23.2
pluggy==1.3.0
pycparser==2.21
pytest==7.4.3
pytest-rerunfailures==13.0
pytest-timeout==2.2.0
pywin32==306
requests==2.31.0
tomli==2.0.1
urllib3==2.1.0
websocket-client==1.6.4
windows-curses==2.3.2
zope.event==5.0
zope.interface==6.1
